Give I32 a typed_native machine representation
I32 already existed — along with I8, I16, U8, U16, U32, U64 — as a
range-checked runtime annotation. `(fn f [x : I32] : I32 ...)` works
interpreted today and rejects both an argument and a return outside 32 bits.
What was missing is the machine representation: resolvedAotRepr knew only I64,
F64 and native pointers, so a typed_native signature could not say `I32` and a
C `int` parameter was unreachable. That is why sqlite_shim.c carried a wrapper
widening the column index to int64_t.

arkI32 is deliberately a *boundary* representation, not an arithmetic one. It
crosses edges — parameter, result, native struct field, FFI argument — and
widens into I64 to be computed with, but arithmetic never produces one:

  (fn read_small   [n : Node] : I32 n/small)        -> int32_t, direct load
  (fn widened      [n : Node] : I64 n/small)        -> widens, fine
  (fn pass_through [n : Node col : I32] : I64 ...)  -> int32_t into C `int`
  (fn wraps        [n : Node] : I32 (+ n/small 1))  -> rejected

The rejection is the point. Gene's I32 is a range-checked Int, so interpreted
`(+ 2147483647 1)` returned as I32 is an error, while C int32_t arithmetic
wraps to -2147483648 without complaint. Mapping I32 onto C arithmetic would
have made compiled and interpreted code disagree — the failure class this
branch has spent its length removing. Computing in I64 keeps them identical,
and narrowing I64 back into I32 stays rejected because a typed_native body has
no error path to report a failed check.

The SQLite example drops a shim function as a result: sqlite3_column_int64 is
bound directly with `[stmt : Stmt column : C/Int]`, the column parameters
become I32 throughout, and out-parameters are now the only reason
sqlite_shim.c exists. Totals are unchanged (340/250/340).

Not covered: a pure scalar I32 function with no native pointer. Those take the
separate scalar AOT path (detectAotExpr), which still knows only I64/F64, so
`^native_entry` on one reports that it requires typed_native lowering. The
other sized annotations (I8/I16/U8/U16/U32/U64) are unrepresented for the same
reason and would follow this same shape.

nimble test, spec (505), perf and wasm pass.

+++ examples/native/main.c
@@ -24,14 +24,14 @@ void gene_sqlite_close(sqlite3 *db);
 int64_t gene_native_step_row(sqlite3_stmt *stmt);
 int64_t gene_native_reset_stmt(sqlite3_stmt *stmt);
 int64_t gene_native_column_count(sqlite3_stmt *stmt);
-int64_t gene_native_column_i64(sqlite3_stmt *stmt, int64_t column);
-int64_t gene_native_read_first(sqlite3_stmt *stmt, int64_t first_column);
-int64_t gene_native_row_total(sqlite3_stmt *stmt, int64_t amount_column,
-                              int64_t quantity_column);
-int64_t gene_native_row_total_capped(sqlite3_stmt *stmt, int64_t amount_column,
-                                     int64_t quantity_column, int64_t cap);
-int64_t gene_native_scan_total(sqlite3_stmt *stmt, int64_t amount_column,
-                               int64_t quantity_column, int64_t row_marker);
+int64_t gene_native_column_i64(sqlite3_stmt *stmt, int32_t column);
+int64_t gene_native_read_first(sqlite3_stmt *stmt, int32_t first_column);
+int64_t gene_native_row_total(sqlite3_stmt *stmt, int32_t amount_column,
+                              int32_t quantity_column);
+int64_t gene_native_row_total_capped(sqlite3_stmt *stmt, int32_t amount_column,
+                                     int32_t quantity_column, int64_t cap);
+int64_t gene_native_scan_total(sqlite3_stmt *stmt, int32_t amount_column,
+                               int32_t quantity_column, int64_t row_marker);
 
 #define ROW 100 /* SQLITE_ROW */
